Part 5 acceptance tests

+++ app/controllers/questions_controller.rb
@@ -1,17 +1,20 @@
 class QuestionsController < ApplicationController
-
+  before_action :authenticate_user!, except: %i[index show]
+  before_action :question, except: %i[create]
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = Answer.new
+  end
 
   def new; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.build(question_params)
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -28,8 +31,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.owner_of?(question)
+      question.destroy 
+      redirect_to questions_path, notice: 'Question deleted'
+    else
+      redirect_to question_path(question), notice: "You can't to that"
+    end
   end
 
   private
